// src/app/core/models/ms-documents/clarification-documents-impro-model.ts
export interface IClarificationDocumentsImpro {
  clarification?: string;
  sender?: string;
  foundation?: string;
  id?: number;
  version?: number;
  transmitterId?: string;
  paragraphInitial?: string;
  applicationId?: number;
  positionSender?: string;
  paragraphFinal?: string;
  consistentIn?: string;
  managedTo?: string;
  invoiceLearned?: string;
  invoiceNumber?: number;
  positionAddressee?: string;
  modificationDate?: Date;
  creationUser?: string;
  documentTypeId?: string;
  modificationUser?: string;
  worthAppraisal?: number;
  creationDate?: Date;
  rejectNoticeId?: number;
  assignmentInvoiceDate?: Date;
  mailNotification?: string;
  areaUserCapture?: string;
}